chore(base): Move MinimalStateEvent higher up the module hierarchy
