+++ crates/matrix-sdk-base/src/rooms/mod.rs
@@ -17,105 +17,13 @@ use ruma::{
             topic::RoomTopicEventContent,
         },
         AnyStrippedStateEvent, AnySyncStateEvent, RedactContent, RedactedEventContent,
-        StateEventContent, StrippedStateEvent, SyncStateEvent,
+        StateEventContent, SyncStateEvent,
     },
-    EventId, OwnedEventId, OwnedUserId, RoomVersionId,
+    EventId, OwnedUserId, RoomVersionId,
 };
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 
-// #[serde(bound)] instead of DeserializeOwned in type where clause does not
-// work, it can only be a single bound that replaces the default and if a helper
-// trait is used, the compiler still complains about Deserialize not being
-// implemented for C::Redacted.
-//
-// It is unclear why a Serialize bound on C::Redacted is not also required.
-#[derive(Clone, Debug, Deserialize, Serialize)]
-enum MinimalStateEvent<C: StateEventContent + RedactContent>
-where
-    C::Redacted: StateEventContent + RedactedEventContent + DeserializeOwned,
-{
-    Original(OriginalMinimalStateEvent<C>),
-    Redacted(RedactedMinimalStateEvent<C::Redacted>),
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-struct OriginalMinimalStateEvent<C>
-where
-    C: StateEventContent,
-{
-    content: C,
-    event_id: Option<OwnedEventId>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-struct RedactedMinimalStateEvent<C>
-where
-    C: StateEventContent + RedactedEventContent,
-{
-    content: C,
-    event_id: Option<OwnedEventId>,
-}
-
-impl<C> MinimalStateEvent<C>
-where
-    C: StateEventContent + RedactContent,
-    C::Redacted: StateEventContent + RedactedEventContent + DeserializeOwned,
-{
-    fn event_id(&self) -> Option<&EventId> {
-        match self {
-            MinimalStateEvent::Original(ev) => ev.event_id.as_deref(),
-            MinimalStateEvent::Redacted(ev) => ev.event_id.as_deref(),
-        }
-    }
-
-    fn as_original(&self) -> Option<&OriginalMinimalStateEvent<C>> {
-        match self {
-            MinimalStateEvent::Original(ev) => Some(ev),
-            MinimalStateEvent::Redacted(_) => None,
-        }
-    }
-
-    fn redact(&mut self, room_version: &RoomVersionId)
-    where
-        C: Clone,
-    {
-        if let MinimalStateEvent::Original(ev) = self {
-            *self = MinimalStateEvent::Redacted(RedactedMinimalStateEvent {
-                content: ev.content.clone().redact(room_version),
-                event_id: ev.event_id.clone(),
-            });
-        }
-    }
-}
-
-impl<C> From<&SyncStateEvent<C>> for MinimalStateEvent<C>
-where
-    C: Clone + StateEventContent + RedactContent,
-    C::Redacted: Clone + StateEventContent + RedactedEventContent + DeserializeOwned,
-{
-    fn from(ev: &SyncStateEvent<C>) -> Self {
-        match ev {
-            SyncStateEvent::Original(ev) => Self::Original(OriginalMinimalStateEvent {
-                content: ev.content.clone(),
-                event_id: Some(ev.event_id.clone()),
-            }),
-            SyncStateEvent::Redacted(ev) => Self::Redacted(RedactedMinimalStateEvent {
-                content: ev.content.clone(),
-                event_id: Some(ev.event_id.clone()),
-            }),
-        }
-    }
-}
-
-impl<C> From<&StrippedStateEvent<C>> for MinimalStateEvent<C>
-where
-    C: Clone + StateEventContent + RedactContent,
-    C::Redacted: StateEventContent + RedactedEventContent + DeserializeOwned,
-{
-    fn from(ev: &StrippedStateEvent<C>) -> Self {
-        Self::Original(OriginalMinimalStateEvent { content: ev.content.clone(), event_id: None })
-    }
-}
+use crate::MinimalStateEvent;
 
 /// The name of the room, either from the metadata or calculated
 /// according to [matrix specification](https://matrix.org/docs/spec/client_server/latest#calculating-the-display-name-for-a-room)
